+++ becpg-core/src/main/java/fr/becpg/repo/product/data/productList/MicrobioListDataItem.java
@@ -17,13 +17,13 @@ public class MicrobioListDataItem {
 	private NodeRef nodeRef;		
 	
 	/** The value. */
-	private Float value;
+	private Double value;
 	
 	/** The unit. */
 	private String unit;
 	
 	/** The maxi. */
-	private Float maxi;
+	private Double maxi;
 	
 	private String textCriteria;
 	
@@ -53,7 +53,7 @@ public class MicrobioListDataItem {
 	 *
 	 * @return the value
 	 */
-	public Float getValue() {
+	public Double getValue() {
 		return value;
 	}
 	
@@ -62,7 +62,7 @@ public class MicrobioListDataItem {
 	 *
 	 * @param value the new value
 	 */
-	public void setValue(Float value) {
+	public void setValue(Double value) {
 		this.value = value;
 	}
 	
@@ -89,7 +89,7 @@ public class MicrobioListDataItem {
 	 *
 	 * @return the maxi
 	 */
-	public Float getMaxi() {
+	public Double getMaxi() {
 		return maxi;
 	}
 	
@@ -98,7 +98,7 @@ public class MicrobioListDataItem {
 	 *
 	 * @param maxi the new maxi
 	 */
-	public void setMaxi(Float maxi) {
+	public void setMaxi(Double maxi) {
 		this.maxi = maxi;
 	}
 	
@@ -144,7 +144,7 @@ public class MicrobioListDataItem {
 	 * @param maxi the maxi
 	 * @param microbio the microbio
 	 */
-	public MicrobioListDataItem(NodeRef nodeRef, Float value, String unit, Float maxi, String textCriteria, NodeRef microbio){
+	public MicrobioListDataItem(NodeRef nodeRef, Double value, String unit, Double maxi, String textCriteria, NodeRef microbio){
 		this.nodeRef = nodeRef;
 		this.value = value;
 		this.unit = unit;
